requirements.txt
mcp>=1.6.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.2 